+++ emg3d/optimize.py
@@ -21,7 +21,6 @@ misfit function and its gradient.
 import numpy as np
 
 from emg3d import maps, fields, utils
-from emg3d.solver import solve
 
 __all__ = ['misfit', 'gradient']
 
@@ -149,6 +148,16 @@ def gradient(simulation):
     The :math:`\partial S`-part takes care of the volume-averaged model
     parameters.
 
+    .. warning::
+
+        To obtain the proper adjoint-state gradient you have to choose linear
+        interpolation for the receiver responses:
+        ``emg3d.Simulation(..., receiver_interpolation='linear')``.
+        The reason is that the point-source is the adjoint of a tri-linear
+        interpolation, so the residual should come from a linear interpolation.
+
+        Also, the adjoint test for magnetic receivers does not yet pass.
+        Electric receivers are good to go.
 
     .. note::