+++ emg3d/simulations.py
@@ -21,7 +21,6 @@ a high-level, specialised modelling tool for the end user.
 # License for the specific language governing permissions and limitations under
 # the License.
 
-import warnings
 import itertools
 from copy import deepcopy
 
@@ -956,6 +955,8 @@ class Simulation:
 
             # Create source.
             if rec.xtype == 'magnetic':
+
+                # TODO: The adjoint test for magnetic receivers does not pass.
                 src_fct = electrodes.TxMagneticDipole
 
                 # If the data is from a magnetic point we have to undo another
@@ -963,7 +964,7 @@ class Simulation:
                 strength /= rfield.smu0
 
             else:
-                src_fct = electrodes.TxElectricDipole
+                src_fct = electrodes.TxElectricPoint
 
             # Get absolute coordinates as fct of source.
             # (Only relevant in case of "relative" receivers.)